Coloring: loosen the conditions to check for existing variables;
The test case exhibits the use of an instruction (rotr) which has a
first input operand that's tied to the one result it produces, and a
second input operand that has a fixed register requirement.

- Because of the fixed register requirement, there's a reassignment
happening from the second operand's current register to the expected
fixed register.
- The fixed register is actually in use by the first operand (fixed
input conflict), so it needs to be diverted and creates its own
solver variable.
- The simple solution doesn't work, and
program_complete_input_constraints will try to create a new solver
variable for the tied input.

However, such a solver variable already exists, and isn't seen by
program_complete_input_constraints (because of the test at the top of
solver::add_live_var), leading to coloring conflicts thereafter.